Moved struct to sinks
